Refactor tests into separate classes
