+++ pi4j-test/src/main/java/com/pi4j/test/About.java
@@ -25,7 +25,6 @@ package com.pi4j.test;
  * #L%
  */
 
-import com.pi4j.context.Context;
 import com.pi4j.exception.Pi4JException;
 import com.pi4j.io.IOType;
 import com.pi4j.platform.Platform;
@@ -43,88 +42,91 @@ public class About {
 
     private static final Logger logger = LoggerFactory.getLogger(About.class);
 
-    /**
-     * <p>Constructor for About.</p>
-     */
-    public About() {
+    private final ProviderContext providerContext;
+
+    public About(ProviderContext providerContext) {
+        this.providerContext = providerContext;
     }
 
     /**
-     * <p>enumerateProviders.</p>
-     *
-     * @param context a {@link com.pi4j.context.Context} object.
+     * Logs and enumerates all providers within the provider context using
+     * a default title.
+     * <p>
+     * This method invokes the overloaded {@code enumerateProviders(String title)}
+     * with the default title, "PROVIDERS". It outputs detailed information about each
+     * provider, including its name, ID, and type. The information is structured
+     * and logged for debugging or analysis purposes.
      */
-    public void enumerateProviders(Context context) {
-        enumerateProviders(context,"PROVIDERS");
+    public void enumerateProviders() {
+        enumerateProviders("PROVIDERS");
     }
+
     /**
-     * <p>enumerateProviders.</p>
+     * Logs and enumerates all providers within the provider context.
+     * This method outputs detailed information about each provider, including its name,
+     * ID, and type. The information is structured and prefixed by a given title.
      *
-     * @param context a {@link com.pi4j.context.Context} object.
-     * @param title a {@link java.lang.String} object.
+     * @param title a String representing the title to be logged before enumerating providers.
      */
-    public void enumerateProviders(Context context, String title) {
+    public void enumerateProviders(String title) {
         logger.info("=====================================================");
         logger.info(title);
         logger.info("=====================================================");
-        for (Provider provider : context.providers().all().values()) {
+        for (Provider provider : providerContext.getContext().providers().all().values()) {
             logger.info("  {} [{}]; {}", provider.name(), provider.id(), provider.type());
         }
     }
 
     /**
-     * <p>enumerateProviders.</p>
+     * Logs and enumerates all providers of the specified IO type within the provider context.
+     * This method retrieves and iterates over all providers of the given IOType from the
+     * provider context, logging their names, IDs, and types. It provides a structured view
+     * of provider information to assist in debugging and analysis of the runtime environment.
      *
-     * @param context a {@link com.pi4j.context.Context} object.
-     * @param ioType a {@link com.pi4j.io.IOType} object.
+     * @param ioType an IOType object representing the type of providers to enumerate.
      */
-    public void enumerateProviders(Context context, IOType ioType) {
+    public void enumerateProviders(IOType ioType) {
         logger.info("=====================================================");
         logger.info("{} PROVIDERS", ioType);
         logger.info("=====================================================");
-        for(var provider : context.providers().all(ioType).values()){
+        for (var provider : providerContext.getContext().providers().all(ioType).values()) {
             logger.info("  {} [{}]; {}", provider.name(), provider.id(), provider.type());
         }
     }
 
     /**
-     * <p>enumeratePlatforms.</p>
-     *
-     * @param context a {@link com.pi4j.context.Context} object.
+     * Logs and enumerates all available platforms within the provider context.
+     * This method retrieves and iterates over all platforms available in the current
+     * runtime environment via the provider context, logging their names, IDs, and descriptions.
+     * It provides a structured display of platform information to facilitate debugging
+     * and runtime environment analysis.
      */
-    public void enumeratePlatforms(Context context) {
+    public void enumeratePlatforms() {
         logger.info("=====================================================");
         logger.info("PLATFORMS");
         logger.info("=====================================================");
-        for (Platform platform : context.platforms().all().values()) {
+        for (Platform platform : providerContext.getContext().platforms().all().values()) {
             logger.info("  {} [{}]; {}", platform.name(), platform.id(), platform.getDescription());
         }
     }
 
     /**
-     * <p>describeDeafultPlatform.</p>
+     * Logs and describes the default platform currently configured in the runtime environment.
+     * This method outputs detailed information about the default platform to the logger and
+     * optionally prints its description to the system output stream. If no default platform
+     * is defined, a warning message is logged.
      *
-     * @param context a {@link com.pi4j.context.Context} object.
-     * @throws Pi4JException if any.
+     * @throws Pi4JException if an error occurs while attempting to describe the default platform.
      */
-    public void describeDefaultPlatform(Context context) throws Pi4JException {
+    public void describeDefaultPlatform() throws Pi4JException {
         logger.info("=====================================================");
         logger.info("DEFAULT (RUNTIME) PLATFORM ");
         logger.info("=====================================================");
-        if (context.platform() == null) {
+        if (providerContext.getContext().platform() == null) {
             logger.warn("  NOT DEFINED");
         } else {
-            logger.info("  {}} [{}}]", context.platform().name(), context.platform().id());
-            context.platform().describe().print(System.out);
+            logger.info("  {}} [{}}]", providerContext.getContext().platform().name(), providerContext.getContext().platform().id());
+            providerContext.getContext().platform().describe().print(System.out);
         }
     }
-
-//    public void enumeratePlatformProviders() throws Pi4JException {
-//        logger.info("=====================================================");
-//        logger.info("PLATFORM PROVIDERS");
-//        logger.info("=====================================================");
-//        for(var provider : Pi4J.platform().providers().values()){
-//            logger.info("  " + provider.name() + "[" + provider.id() + "]; " + provider.type());
-//        }
-//    }
 }
