--- pi4j-test/src/main/java/com/pi4j/test/ContextUtil.java
@@ -1,34 +1,0 @@
-/*
- *
- *
- *  #%L
- *  **********************************************************************
- *  ORGANIZATION  :  Pi4J
- *  PROJECT       :  Pi4J :: LIBRARY  :: Java Library (CORE)
- *  FILENAME      :  ProviderHwTest.java
- *
- *  This file is part of the Pi4J project. More information about
- *  this project can be found here:  https://pi4j.com/
- *  **********************************************************************
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- *  #L%
- *
- *
- */
-
-package com.pi4j.test;
-
-public class ContextUtil {
-
-}
