+++ pi4j-test/src/main/java/com/pi4j/test/ProviderContext.java
@@ -1,0 +1,126 @@
+package com.pi4j.test;
+
+import com.pi4j.Pi4J;
+import com.pi4j.context.Context;
+import com.pi4j.plugin.ffm.providers.gpio.DigitalInputFFMProviderImpl;
+import com.pi4j.plugin.ffm.providers.gpio.DigitalOutputFFMProviderImpl;
+import com.pi4j.plugin.ffm.providers.i2c.I2CFFMProviderImpl;
+import com.pi4j.plugin.ffm.providers.pwm.PwmFFMProviderImpl;
+import com.pi4j.plugin.ffm.providers.serial.SerialFFMProviderImpl;
+import com.pi4j.plugin.ffm.providers.spi.SpiFFMProviderImpl;
+import com.pi4j.plugin.gpiod.provider.gpio.digital.GpioDDigitalInputProvider;
+import com.pi4j.plugin.gpiod.provider.gpio.digital.GpioDDigitalOutputProvider;
+import com.pi4j.plugin.linuxfs.provider.i2c.LinuxFsI2CProvider;
+import com.pi4j.plugin.linuxfs.provider.pwm.LinuxFsPwmProvider;
+import com.pi4j.plugin.linuxfs.provider.spi.LinuxFsSpiProvider;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class ProviderContext {
+
+    private static final Logger logger = LoggerFactory.getLogger(ProviderContext.class);
+
+    public enum TestProvider {
+        FFM,
+        LINUXFS,
+        UNDEFINED;
+
+        public static TestProvider getByName(String name) {
+            for (TestProvider provider : TestProvider.values()) {
+                if (provider.name().equalsIgnoreCase(name)) {
+                    return provider;
+                }
+            }
+            return UNDEFINED;
+        }
+    }
+
+    private final TestProvider testProvider;
+
+    private Context pi4j = null;
+
+    private String i2cProviderName = "";
+    private String spiProviderName = "";
+    private String pwmProviderName = "";
+    private String digitalOutputProviderName = "";
+    private String digitalInputProviderName = "";
+    private String serialProviderName = "";
+
+    public static String DEFAULT_PWM_FILESYSTEM_PATH = "/sys/class/pwm";
+
+    /**
+     *
+     * @param testProvider Identifies which set of providers to create
+     */
+    ProviderContext(TestProvider testProvider) {
+        this.testProvider = testProvider;
+
+        switch (testProvider) {
+            case LINUXFS -> {
+                String pwmFileSystemPath = DEFAULT_PWM_FILESYSTEM_PATH;
+                pi4j = Pi4J.newContextBuilder().add(LinuxFsI2CProvider.newInstance())
+                    .add(GpioDDigitalInputProvider.newInstance())
+                    .add(GpioDDigitalOutputProvider.newInstance())
+                    .add(LinuxFsPwmProvider.newInstance(pwmFileSystemPath))
+                    .add(LinuxFsI2CProvider.newInstance())
+                    .add(LinuxFsSpiProvider.newInstance())
+                    .build();
+                i2cProviderName = "linuxfs-i2c";
+                spiProviderName = "linuxfs-spi";
+                pwmProviderName = "linuxfs-pwm";
+                digitalOutputProviderName = "gpiod-digital-output";
+                digitalInputProviderName = "gpiod-digital-input";
+                serialProviderName = "NONE-serial";
+            }
+            case FFM -> {
+                pi4j = Pi4J.newContextBuilder()
+                    .add(new DigitalOutputFFMProviderImpl())
+                    .add(new DigitalInputFFMProviderImpl())
+                    .add(new I2CFFMProviderImpl())
+                    .add(new SpiFFMProviderImpl())
+                    .add(new PwmFFMProviderImpl())
+                    .add(new SerialFFMProviderImpl())
+                    .build();
+                i2cProviderName = "ffm-i2c";
+                spiProviderName = "ffm-spi";
+                pwmProviderName = "ffm-pwm";
+                digitalOutputProviderName = "ffm-digital-output";
+                digitalInputProviderName = "ffm-digital-input";
+                serialProviderName = "ffm-serial";
+            }
+            default -> logger.error("No test provider specified");
+        }
+    }
+
+    public TestProvider getTestProvider() {
+        return testProvider;
+    }
+
+    public Context getContext() {
+        return pi4j;
+    }
+
+    public String getI2cProviderName() {
+        return i2cProviderName;
+    }
+
+    public String getSpiProviderName() {
+        return spiProviderName;
+    }
+
+    public String getPwmProviderName() {
+        return pwmProviderName;
+    }
+
+    public String getDigitalOutputProviderName() {
+        return digitalOutputProviderName;
+    }
+
+    public String getDigitalInputProviderName() {
+        return digitalInputProviderName;
+    }
+
+    public String getSerialProviderName() {
+        return serialProviderName;
+    }
+}
