+++ pi4j-test/src/main/java/com/pi4j/test/TestResult.java
@@ -1,0 +1,10 @@
+package com.pi4j.test;
+
+public record TestResult(String name, boolean success, String message) {
+    public String getLogOutput() {
+        return "Test " + name + "\n"
+            + (success ? "PASSED" : "FAILED") + "\n"
+            + message + "\n"
+            + "==============================\n";
+    }
+}
