--- pi4j-test/src/main/java/com/pi4j/test/provider/TestDigitalInput.java
@@ -1,76 +1,0 @@
-package com.pi4j.test.provider;
-
-/*-
- * #%L
- * **********************************************************************
- * ORGANIZATION  :  Pi4J
- * PROJECT       :  Pi4J :: TESTING  :: Unit/Integration Tests
- * FILENAME      :  TestDigitalInput.java
- *
- * This file is part of the Pi4J project. More information about
- * this project can be found here:  https://pi4j.com/
- * **********************************************************************
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * #L%
- */
-
-import com.pi4j.io.gpio.digital.*;
-
-/**
- * <p>TestDigitalInput class.</p>
- *
- * @author Robert Savage (<a href="http://www.savagehomeautomation.com">http://www.savagehomeautomation.com</a>)
- * @version $Id: $Id
- */
-public class TestDigitalInput extends DigitalInputBase implements DigitalInput {
-
-    private DigitalState state = DigitalState.UNKNOWN;
-
-    /** {@inheritDoc} */
-    @Override
-    public DigitalState state() {
-        return this.state;
-    }
-
-    /**
-     * <p>Constructor for TestDigitalInput.</p>
-     *
-     * @param provider a {@link com.pi4j.io.gpio.digital.DigitalInputProvider} object.
-     * @param config a {@link com.pi4j.io.gpio.digital.DigitalInputConfig} object.
-     */
-    public TestDigitalInput(DigitalInputProvider provider, DigitalInputConfig config){
-        super(provider, config);
-    }
-
-    /**
-     * <p>test.</p>
-     *
-     * @param state a {@link com.pi4j.io.gpio.digital.DigitalState} object.
-     * @return a {@link com.pi4j.test.provider.TestDigitalInput} object.
-     */
-    public TestDigitalInput test(DigitalState state){
-
-        // check to see of there is a state change; if there is then we need
-        // to update the internal value variable and dispatch the change event
-        if(this.state().equals(state)) {
-
-            // update current/new value
-            this.state = state;
-
-            // dispatch value change event
-            this.dispatch(new DigitalStateChangeEvent(this, this.state()));
-        }
-        return this;
-    }
-}
